commit automatique

+++ src/controller/views.py
@@ -56,7 +56,9 @@ def info(request, resource, name):
     return render(request, "controller/info.html", context={"info": separated, "resource_type": resource_type, "name": name})
 
 def restart(request, resource, name):
-    api.command(f"kubectl delete pod { name }")
+    print(resource)
+    if resource == "pod":
+        api.command(f"kubectl delete pod { name }")
     namelist = api.command("kubectl get pods --output name")
     number_of_pods = api.command("kubectl get pods --output name | wc -l")
     separated = namelist.split("\n")
